pyproject.toml
[tool.poetry]
name = "gh-automation-base"
version = "0.1.0"
description = "A project base for automating Python tasks using GitHub actions and Google tools."
authors = ["Samuel Giffard <samuel@giffard.co>"]
license = "MIT"
readme = "README.md"
packages = [{include = "gh_automation_base"}]

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.9.0"
pydantic = "^2.2.1"
pydantic-settings = "^2.0.3"
python-dotenv = "^1.0.0"
asyncpg = "^0.28.0"

[tool.poetry.scripts]
auto = "gh_automation_base.cli:main"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.275"
black = "^23.3.0"
mypy = "^1.4.0"
pre-commit = "^3.3.3"
yoyo-migrations = "^8.2.0"
psycopg = "^3.1.10"
asyncpg-stubs = "^0.28.0"

[tool.ruff]
target-version = "py311"

[tool.ruff.mccabe]
max-complexity = 10

[tool.mypy]
strict = "True"
plugins = "pydantic.mypy"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
